requirements.txt
tensorflow
faiss-cpu